refactor: consolidate sqlite queries and prune storage drafts
